fix(android): unblock release builds and CI on API 37
Use compileSdk minorApiLevel 0 for android-37.0, enable BuildConfig, read version.properties, and fix poll/ZXing/R8 issues.

+++ android/app/src/main/kotlin/app/zerorelay/data/chat/ChatRepository.kt
@@ -338,7 +338,7 @@ class ChatRepository(
                 try {
                     val token = roomAccessToken ?: break
                     val timeoutMs = ProtocolPolicy.POLL_TIMEOUT_MS
-                    val frame = RelayTunnel.encodePoll(token, lastPollTimestamp, timeoutMs)
+                    val frame = RelayTunnel.encodePoll(token, lastPollTimestamp, timeoutMs.toInt())
                     val request = Request.Builder()
                         .url("$base/t")
                         .post(frame.toRequestBody("application/octet-stream".toMediaType()))
